requirements.txt
asgiref==3.9.1
Django==4.2.23
django-monthfield==0.1.4
sqlparse==0.5.3
typing_extensions==4.14.1
uuid==1.30
